+++ crates/nu-command/src/core_commands/overlay/mod.rs
@@ -1,11 +1,11 @@
-mod add;
 mod command;
 mod hide;
 mod list;
 mod new;
+mod use_;
 
-pub use add::OverlayUse;
 pub use command::Overlay;
 pub use hide::OverlayHide;
 pub use list::OverlayList;
 pub use new::OverlayNew;
+pub use use_::OverlayUse;
